Module tauri::api::dialog::blocking

Blocking interfaces for the dialog APIs.

The blocking APIs will block the current thread to execute instead of relying on callback closures, which makes them easier to use.

NOTE: You cannot block the main thread when executing the dialog APIs, so you must use the crate::api::dialog methods instead. Examples of main thread context are the crate::App::run closure and non-async commmands.

Structs

FileDialogBuilder

The file dialog builder.

MessageDialogBuilder

A builder for message dialogs.

Functions

ask

Displays a dialog with a message and an optional title with a “yes” and a “no” button and wait for it to be closed.

confirm

Displays a dialog with a message and an optional title with an “ok” and a “cancel” button and wait for it to be closed.

message

Displays a message dialog and wait for it to be closed.