Module tauri::api::ipc

Types and functions related to Inter Procedure Call(IPC).

This module includes utilities to send messages to the JS layer of the webview.

Structs

CallbackFn

The Callback type is the return value of the transformCallback JavaScript function.

SerializeOptions

Optional settings to pass to the templating system.

Functions

format_callback

Formats a function name and argument to be evaluated as callback.

format_callback_result

Formats a Result type to its Promise response. Useful for Promises handling. If the Result is_ok(), the callback will be the success_callback function name and the argument will be the Ok value. If the Result is_err(), the callback will be the error_callback function name and the argument will be the Err value.

serialize_js

Transforms & escapes a JSON value.

serialize_js_with

Transforms & escapes a JSON value.