Enum tauri::api::path::BaseDirectory

#[repr(u16)]
#[non_exhaustive]
pub enum BaseDirectory {

    Audio,
    Cache,
    Config,
    Data,
    LocalData,
    Desktop,
    Document,
    Download,
    Executable,
    Font,
    Home,
    Picture,
    Public,
    Runtime,
    Template,
    Video,
    Resource,
    App,
    Log,
    Temp,
    AppConfig,
    AppData,
    AppLocalData,
    AppCache,
    AppLog,

}

A base directory to be used in resolve_path.

The base directory is the optional root of a file system operation. If informed by the API call, all paths will be relative to the path of the given directory.

For more information, check the dirs_next documentation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Audio

The Audio directory.

Cache

The Cache directory.

Config

The Config directory.

Data

The Data directory.

LocalData

The LocalData directory.

Desktop

The Desktop directory.

Document

The Document directory.

Download

The Download directory.

Executable

The Executable directory.

Font

The Font directory.

Home

The Home directory.

Picture

The Picture directory.

Public

The Public directory.

Runtime

The Runtime directory.

Template

The Template directory.

Video

The Video directory.

Resource

The Resource directory.

App

๐Deprecated since 1.2.0: Will be removed in 2.0.0. Use BaseDirectory::AppConfig or BaseDirectory::AppData` instead.

The default app config directory. Resolves to BaseDirectory::Config/{bundle_identifier}.

Log

๐Deprecated since 1.2.0: Will be removed in 2.0.0. Use BaseDirectory::AppLog instead.

The default app log directory. Resolves to BaseDirectory::Home/Library/Logs/{bundle_identifier} on macOS and BaseDirectory::Config/{bundle_identifier}/logs on linux and Windows.

Temp

A temporary directory. Resolves to temp_dir.

AppConfig

The default app config directory. Resolves to BaseDirectory::Config/{bundle_identifier}.

AppData

The default app data directory. Resolves to BaseDirectory::Data/{bundle_identifier}.

AppLocalData

The default app local data directory. Resolves to BaseDirectory::LocalData/{bundle_identifier}.

AppCache

The default app cache directory. Resolves to BaseDirectory::Cache/{bundle_identifier}.

AppLog

The default app log directory. Resolves to BaseDirectory::Home/Library/Logs/{bundle_identifier} on macOS and BaseDirectory::Config/{bundle_identifier}/logs on linux and Windows.

Implementations

impl BaseDirectory

pub fn variable(self) -> &'static str

Gets the variable that represents this BaseDirectory for string paths.

pub fn from_variable(variable: &str) -> Option<Self>

Gets the BaseDirectory associated with the given variable, or None if the variable doesnโt match any.