Enum tauri::Error

#[non_exhaustive]
pub enum Error {

    Runtime(Error),
    CreateWindow,
    WindowLabelAlreadyExists(String),
    WebviewNotFound,
    FailedToSendMessage,
    AssetNotFound(String),
    Json(Error),
    UnknownApi(Option<Error>),
    FailedToExecuteApi(Error),
    Io(Error),
    InvalidIcon(Error),
    HttpClientNotInitialized,
    ApiNotAllowlisted(String),
    InvalidArgs(&'static str, &'static str, Error),
    Setup(SetupError),
    PluginInitialization(String, String),
    InvalidUrl(ParseError),
    JoinError(JoinError),
    PathNotAllowed(PathBuf),
    NotificationNotAllowed,
    UrlNotAllowed(Url),
    SidecarNotAllowed(PathBuf),
    SidecarNotFound(Box<ShellScopeError>),
    ProgramNotAllowed(PathBuf),
    InvalidWindowUrl(&'static str),
    GlobPattern(PatternError),
    InvalidWindowHandle,

}

Runtime errors that can happen inside a Tauri application.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Runtime(Error)

Runtime error.

CreateWindow

Failed to create window.

WindowLabelAlreadyExists(String)

Window label must be unique.

WebviewNotFound

Can’t access webview dispatcher because the webview was closed or not found.

FailedToSendMessage

Failed to send message to webview.

AssetNotFound(String)

Embedded asset not found.

Json(Error)

Failed to serialize/deserialize.

UnknownApi(Option<Error>)

Unknown API type.

FailedToExecuteApi(Error)

Failed to execute tauri API.

Io(Error)

IO error.

InvalidIcon(Error)

Failed to load window icon.

HttpClientNotInitialized

Client with specified ID not found.

ApiNotAllowlisted(String)

API not whitelisted on tauri.conf.json

InvalidArgs(&'static str, &'static str, Error)

Invalid args when running a command.

Setup(SetupError)

Encountered an error in the setup hook,

PluginInitialization(String, String)

Error initializing plugin.

InvalidUrl(ParseError)

A part of the URL is malformed or invalid. This may occur when parsing and combining user-provided URLs and paths.

JoinError(JoinError)

Task join error.

PathNotAllowed(PathBuf)

Path not allowed by the scope.

NotificationNotAllowed

The user did not allow sending notifications.

UrlNotAllowed(Url)

URL not allowed by the scope.

SidecarNotAllowed(PathBuf)

Sidecar not allowed by the configuration.

SidecarNotFound(Box<ShellScopeError>)

Sidecar was not found by the configuration.

ProgramNotAllowed(PathBuf)

Program not allowed by the scope.

InvalidWindowUrl(&'static str)

An invalid window URL was provided. Includes details about the error.

GlobPattern(PatternError)

Invalid glob pattern.

InvalidWindowHandle

The Window’s raw handle is invalid for the platform.