Enum tauri::RunEvent

#[non_exhaustive]
pub enum RunEvent {
    Exit,
    ExitRequested {
        api: ExitRequestApi,
    },
    WindowEvent {
        label: String,
        event: WindowEvent,
    },
    Ready,
    Resumed,
    MainEventsCleared,
}

An application event, triggered from the event loop.

See App::run for usage examples.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Exit

Event loop is exiting.

ExitRequested

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

api: ExitRequestApi

Event API

The app is about to exit

WindowEvent

Fields

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

label: String

The window label.

event: WindowEvent

The detailed event.

An event associated with a window.

Ready

Application ready.

Resumed

Sent if the event loop is being resumed.

MainEventsCleared

Emitted when all of the event loop’s input events have been processed and redraw processing is about to begin.

This event is useful as a place to put your code that should be run after all state-changing events have been handled and you want to do stuff (updating state, performing calculations, etc) that happens as the “main body” of your event loop.