Module tauri::ipc

Types and functions related to Inter Procedure Call(IPC).

This module includes utilities to send messages to the JS layer of the webview.

Structs

CallbackFn

The Callback type is the return value of the transformCallback JavaScript function.

CapabilityBuilder

A builder for a Capability.

Channel

An IPC channel.

CommandItem

Represents a custom command.

CommandScope

Access scope for a command that can be retrieved directly in the command function.

GlobalScope

Global access scope that can be retrieved directly in the command function.

Invoke

The message and resolver given to a custom command.

InvokeError

Error response from an InvokeMessage.

InvokeMessage

An invoke message.

InvokeResolver

Resolver of a invoke message.

JavaScriptChannelId

The ID of a channel that was defined on the JavaScript layer.

Request

The IPC request.

Response

The IPC response.

RuntimeAuthority

The runtime authority used to authorize IPC execution based on the Access Control List.

ScopeValue

List of allowed and denied objects that match either the command-specific or plugin global scope criteria.

SerializeOptions

Optional settings to pass to the templating system.

Enums

InvokeBody

Possible values of an IPC payload.

InvokeResponse

Response from a InvokeMessage passed to the InvokeResolver.

InvokeResponseBody

Possible values of an IPC response.

Origin

The origin trying to access the IPC.

Traits

CommandArg

Trait implemented by command arguments to derive a value from a CommandItem.

IpcResponse

Marks a type as a response to an IPC call.

RuntimeCapability

A capability that can be added at runtime.

ScopeObject

Marks a type as a scope object.

ScopeObjectMatch

A ScopeObject whose validation can be represented as a bool.

Type Aliases

InvokeHandler

A closure that is run every time Tauri receives a message it doesn’t explicitly handle.

InvokeResponder

A closure that is responsible for respond a JS message.

OwnedInvokeResponder

Similar to InvokeResponder but taking owned arguments.