Trait tauri::ipc::ScopeObject

pub trait ScopeObject:
    Sized
    + Send
    + Sync
    + Debug
    + 'static {
    type Error: Error + Send + Sync;

    // Required method
    fn deserialize<R: Runtime>(
        app: &AppHandle<R>,
        raw: Value,
    ) -> Result<Self, Self::Error>;
}

Marks a type as a scope object.

Usually you will just rely on serde::de::DeserializeOwned instead of implementing it manually, though this is useful if you need to do some initialization logic on the type itself.

Required Associated Types

type Error: Error + Send + Sync

The error type.

Required Methods

fn deserialize<R: Runtime>( app: &AppHandle<R>, raw: Value, ) -> Result<Self, Self::Error>

Deserialize the raw scope value.

Object Safety

This trait is not object safe.