Module tauri::menu

Menu types and utilities.

Structs

AboutMetadata

Application metadata for the PredefinedMenuItem::about.

AboutMetadataBuilder

A builder type for AboutMetadata.

CheckMenuItem

A menu item inside a Menu or Submenu and usually contains a text and a check mark or a similar toggle that corresponds to a checked and unchecked states.

CheckMenuItemBuilder

A builder type for CheckMenuItem

IconMenuItem

A menu item inside a Menu or Submenu and usually contains an icon and a text.

IconMenuItemBuilder

A builder type for IconMenuItem

Menu

A type that is either a menu bar on the window on Windows and Linux or as a global menu in the menubar on macOS.

MenuBuilder

A builder type for Menu

MenuEvent

Describes a menu event emitted when a menu item is activated

MenuId

An unique id that is associated with a menu or a menu item.

MenuItem

A menu item inside a Menu or Submenu and contains only text.

MenuItemBuilder

A builder type for MenuItem

PredefinedMenuItem

A predefined (native) menu item which has a predefined behavior by the OS or by this crate.

Submenu

A type that is a submenu inside a Menu or Submenu

SubmenuBuilder

A builder type for Submenu

Enums

MenuItemKind

An enumeration of all menu item kinds that could be added to a Menu or Submenu

NativeIcon

A native Icon to be used for the menu item

Constants

HELP_SUBMENU_ID

Expected submenu id of the Help menu for macOS.

WINDOW_SUBMENU_ID

Expected submenu id of the Window menu for macOS.

Traits

ContextMenu

A helper trait with methods to help creating a context menu.

IsMenuItem

A trait that defines a generic item in a menu, which may be one of MenuItemKind