Struct tauri::menu::Menu

pub struct Menu<R: Runtime>(/* private fields */);

A type that is either a menu bar on the window on Windows and Linux or as a global menu in the menubar on macOS.

Implementations

impl<R: Runtime> Menu<R>

pub fn new<M: Manager<R>>(manager: &M) -> Result<Self>

Creates a new menu.

pub fn with_id<M: Manager<R>, I: Into<MenuId>>( manager: &M, id: I, ) -> Result<Self>

Creates a new menu with the specified id.

pub fn with_items<M: Manager<R>>( manager: &M, items: &[&dyn IsMenuItem<R>], ) -> Result<Self>

Creates a new menu with given items. It calls Menu::new and Menu::append_items internally.

pub fn with_id_and_items<M: Manager<R>, I: Into<MenuId>>( manager: &M, id: I, items: &[&dyn IsMenuItem<R>], ) -> Result<Self>

Creates a new menu with the specified id and given items. It calls Menu::new and Menu::append_items internally.

pub fn default(app_handle: &AppHandle<R>) -> Result<Self>

Creates a menu filled with default menu items and submenus.

pub fn app_handle(&self) -> &AppHandle<R>

The application handle associated with this type.

pub fn id(&self) -> &MenuId

Returns a unique identifier associated with this menu.

pub fn append(&self, item: &dyn IsMenuItem<R>) -> Result<()>

Add a menu item to the end of this menu.

Platform-specific:

macOS: Only Submenu can be added to the menu.

pub fn append_items(&self, items: &[&dyn IsMenuItem<R>]) -> Result<()>

Add menu items to the end of this menu. It calls Menu::append in a loop internally.

Platform-specific:

macOS: Only Submenu can be added to the menu

pub fn prepend(&self, item: &dyn IsMenuItem<R>) -> Result<()>

Add a menu item to the beginning of this menu.

Platform-specific:

macOS: Only Submenu can be added to the menu

pub fn prepend_items(&self, items: &[&dyn IsMenuItem<R>]) -> Result<()>

Add menu items to the beginning of this menu. It calls Menu::insert_items with position of 0 internally.

Platform-specific:

macOS: Only Submenu can be added to the menu

pub fn insert(&self, item: &dyn IsMenuItem<R>, position: usize) -> Result<()>

Insert a menu item at the specified position in the menu.

Platform-specific:

macOS: Only Submenu can be added to the menu

pub fn insert_items( &self, items: &[&dyn IsMenuItem<R>], position: usize, ) -> Result<()>

Insert menu items at the specified position in the menu.

Platform-specific:

macOS: Only Submenu can be added to the menu

pub fn remove(&self, item: &dyn IsMenuItem<R>) -> Result<()>

Remove a menu item from this menu.

pub fn remove_at(&self, position: usize) -> Result<Option<MenuItemKind<R>>>

Remove the menu item at the specified position from this menu and returns it.

pub fn get<'a, I>(&self, id: &'a I) -> Option<MenuItemKind<R>>
where I: ?Sized, MenuId: PartialEq<&'a I>,

Retrieves the menu item matching the given identifier.

pub fn items(&self) -> Result<Vec<MenuItemKind<R>>>

Returns a list of menu items that has been added to this menu.

pub fn set_as_app_menu(&self) -> Result<Option<Menu<R>>>

Set this menu as the application menu.

This is an alias for AppHandle::set_menu.

pub fn set_as_window_menu(&self, window: &Window<R>) -> Result<Option<Menu<R>>>

Set this menu as the window menu.

This is an alias for Window::set_menu.