Struct tauri::menu::MenuItemBuilder

pub struct MenuItemBuilder { /* private fields */ }

A builder type for MenuItem

Implementations

impl MenuItemBuilder

pub fn new<S: AsRef<str>>(text: S) -> Self

Create a new menu item builder.

text could optionally contain an & before a character to assign this character as the mnemonic for this menu item. To display a & without assigning a mnemenonic, use &&.

pub fn with_id<I: Into<MenuId>, S: AsRef<str>>(id: I, text: S) -> Self

Create a new menu item builder with the specified id.

text could optionally contain an & before a character to assign this character as the mnemonic for this menu item. To display a & without assigning a mnemenonic, use &&.

pub fn id<I: Into<MenuId>>(self, id: I) -> Self

Set the id for this menu item.

pub fn enabled(self, enabled: bool) -> Self

Set the enabled state for this menu item.

pub fn accelerator<S: AsRef<str>>(self, accelerator: S) -> Self

Set the accelerator for this menu item.

pub fn build<R: Runtime, M: Manager<R>>( self, manager: &M, ) -> Result<MenuItem<R>>

Build the menu item