Struct tauri::menu::Submenu

pub struct Submenu<R: Runtime>(/* private fields */);

A type that is a submenu inside a Menu or Submenu

Implementations

impl<R: Runtime> Submenu<R>

pub fn new<M: Manager<R>, S: AsRef<str>>( manager: &M, text: S, enabled: bool, ) -> Result<Self>

Creates a new submenu.

pub fn with_id<M: Manager<R>, I: Into<MenuId>, S: AsRef<str>>( manager: &M, id: I, text: S, enabled: bool, ) -> Result<Self>

Creates a new submenu with the specified id.

pub fn with_items<M: Manager<R>, S: AsRef<str>>( manager: &M, text: S, enabled: bool, items: &[&dyn IsMenuItem<R>], ) -> Result<Self>

Creates a new menu with given items. It calls Submenu::new and Submenu::append_items internally.

pub fn with_id_and_items<M: Manager<R>, I: Into<MenuId>, S: AsRef<str>>( manager: &M, id: I, text: S, enabled: bool, items: &[&dyn IsMenuItem<R>], ) -> Result<Self>

Creates a new menu with the specified id and given items. It calls Submenu::new and Submenu::append_items internally.

pub fn app_handle(&self) -> &AppHandle<R>

The application handle associated with this type.

pub fn id(&self) -> &MenuId

Returns a unique identifier associated with this submenu.

pub fn append(&self, item: &dyn IsMenuItem<R>) -> Result<()>

Add a menu item to the end of this submenu.

pub fn append_items(&self, items: &[&dyn IsMenuItem<R>]) -> Result<()>

Add menu items to the end of this submenu. It calls Submenu::append in a loop internally.

pub fn prepend(&self, item: &dyn IsMenuItem<R>) -> Result<()>

Add a menu item to the beginning of this submenu.

pub fn prepend_items(&self, items: &[&dyn IsMenuItem<R>]) -> Result<()>

Add menu items to the beginning of this submenu. It calls Submenu::insert_items with position of 0 internally.

pub fn insert(&self, item: &dyn IsMenuItem<R>, position: usize) -> Result<()>

Insert a menu item at the specified position in this submenu.

pub fn insert_items( &self, items: &[&dyn IsMenuItem<R>], position: usize, ) -> Result<()>

Insert menu items at the specified position in this submenu.

pub fn remove(&self, item: &dyn IsMenuItem<R>) -> Result<()>

Remove a menu item from this submenu.

pub fn remove_at(&self, position: usize) -> Result<Option<MenuItemKind<R>>>

Remove the menu item at the specified position from this submenu and returns it.

pub fn get<'a, I>(&self, id: &'a I) -> Option<MenuItemKind<R>>
where I: ?Sized, MenuId: PartialEq<&'a I>,

Retrieves the menu item matching the given identifier.

pub fn items(&self) -> Result<Vec<MenuItemKind<R>>>

Returns a list of menu items that has been added to this submenu.

pub fn text(&self) -> Result<String>

Get the text for this submenu.

pub fn set_text<S: AsRef<str>>(&self, text: S) -> Result<()>

Set the text for this submenu. text could optionally contain an & before a character to assign this character as the mnemonic for this submenu. To display a & without assigning a mnemonic, use &&.

pub fn is_enabled(&self) -> Result<bool>

Get whether this submenu is enabled or not.

pub fn set_enabled(&self, enabled: bool) -> Result<()>

Enable or disable this submenu.

pub fn set_as_windows_menu_for_nsapp(&self) -> Result<()>

Set this submenu as the Window menu for the application on macOS.

This will cause macOS to automatically add window-switching items and certain other items to the menu.

pub fn set_as_help_menu_for_nsapp(&self) -> Result<()>

Set this submenu as the Help menu for the application on macOS.

This will cause macOS to automatically add a search box to the menu.

If no menu is set as the Help menu, macOS will automatically use any menu which has a title matching the localized word โHelpโ.