Enum tauri::path::BaseDirectory

#[non_exhaustive]

#[repr(u16)]
pub enum BaseDirectory {

    Audio = 1,
    Cache = 2,
    Config = 3,
    Data = 4,
    LocalData = 5,
    Document = 6,
    Download = 7,
    Picture = 8,
    Public = 9,
    Video = 10,
    Resource = 11,
    Temp = 12,
    AppConfig = 13,
    AppData = 14,
    AppLocalData = 15,
    AppCache = 16,
    AppLog = 17,
    Desktop = 18,
    Executable = 19,
    Font = 20,
    Home = 21,
    Runtime = 22,
    Template = 23,

}

A base directory for a path.

The base directory is the optional root of a file system operation. If informed by the API call, all paths will be relative to the path of the given directory.

For more information, check the dirs documentation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Audio = 1

The Audio directory.

Cache = 2

The Cache directory.

Config = 3

The Config directory.

Data = 4

The Data directory.

LocalData = 5

The LocalData directory.

Document = 6

The Document directory.

Download = 7

The Download directory.

Picture = 8

The Picture directory.

Public = 9

The Public directory.

Video = 10

The Video directory.

Resource = 11

The Resource directory.

Temp = 12

A temporary directory. Resolves to std::env::temp_dir.

AppConfig = 13

The default app config directory. Resolves to BaseDirectory::Config/{bundle_identifier}.

AppData = 14

The default app data directory. Resolves to BaseDirectory::Data/{bundle_identifier}.

AppLocalData = 15

The default app local data directory. Resolves to BaseDirectory::LocalData/{bundle_identifier}.

AppCache = 16

The default app cache directory. Resolves to BaseDirectory::Cache/{bundle_identifier}.

AppLog = 17

The default app log directory. Resolves to BaseDirectory::Home/Library/Logs/{bundle_identifier} on macOS and BaseDirectory::Config/{bundle_identifier}/logs on linux and Windows.

Desktop = 18

The Desktop directory.

Executable = 19

The Executable directory.

Font = 20

The Font directory.

Home = 21

The Home directory.

Runtime = 22

The Runtime directory.

Template = 23

The Template directory.

Implementations

impl BaseDirectory

pub fn variable(self) -> &'static str

Gets the variable that represents this BaseDirectory for string paths.

pub fn from_variable(variable: &str) -> Option<Self>

Gets the BaseDirectory associated with the given variable, or None if the variable doesnโt match any.