Struct tauri::scope::fs::Scope

pub struct Scope { /* private fields */ }

Scope for filesystem access.

Implementations

impl Scope

pub fn new<R: Runtime, M: Manager<R>>( manager: &M, scope: &FsScope, ) -> Result<Self>

Creates a new scope from a FsScope configuration.

pub fn allowed_patterns(&self) -> HashSet<Pattern>

The list of allowed patterns.

pub fn forbidden_patterns(&self) -> HashSet<Pattern>

The list of forbidden patterns.

pub fn listen<F: Fn(&Event) + Send + 'static>(&self, f: F) -> ScopeEventId

Listen to an event on this scope.

pub fn once<F: FnOnce(&Event) + Send + 'static>(&self, f: F) -> ScopeEventId

Listen to an event on this scope and immediately unlisten.

pub fn unlisten(&self, id: ScopeEventId)

Removes an event listener on this scope.

pub fn allow_directory<P: AsRef<Path>>( &self, path: P, recursive: bool, ) -> Result<()>

Extend the allowed patterns with the given directory.

After this function has been called, the frontend will be able to use the Tauri API to read the directory and all of its files. If recursive is true, subdirectories will be accessible too.

pub fn allow_file<P: AsRef<Path>>(&self, path: P) -> Result<()>

Extend the allowed patterns with the given file path.

After this function has been called, the frontend will be able to use the Tauri API to read the contents of this file.

pub fn forbid_directory<P: AsRef<Path>>( &self, path: P, recursive: bool, ) -> Result<()>

Set the given directory path to be forbidden by this scope.

Note: this takes precedence over allowed paths, so its access gets denied always.

pub fn forbid_file<P: AsRef<Path>>(&self, path: P) -> Result<()>

Set the given file path to be forbidden by this scope.

Note: this takes precedence over allowed paths, so its access gets denied always.

pub fn is_allowed<P: AsRef<Path>>(&self, path: P) -> bool

Determines if the given path is allowed on this scope.