Struct tauri::App

pub struct App<R: Runtime = Wry> { /* private fields */ }

The instance of the currently running application.

This type implements Manager which allows for manipulation of global application items.

Implementations

impl App<Wry>

APIs specific to the wry runtime.

pub fn wry_plugin<P: PluginBuilder<EventLoopMessage> + Send + 'static>( &mut self, plugin: P, )
where <P as PluginBuilder<EventLoopMessage>>::Plugin: Send,

Adds a tauri_runtime_wry::Plugin using its tauri_runtime_wry::PluginBuilder.

Stability

This API is unstable.

impl<R: Runtime> App<R>

pub fn on_menu_event<F: Fn(&AppHandle<R>, MenuEvent) + Send + Sync + 'static>( &self, handler: F, )

Registers a global menu event listener.

pub fn config(&self) -> &Config

Gets the app’s configuration, defined on the tauri.conf.json file.

pub fn package_info(&self) -> &PackageInfo

Gets the app’s package information.

pub fn asset_resolver(&self) -> AssetResolver<R>

The application’s asset resolver.

pub fn primary_monitor(&self) -> Result<Option<Monitor>>

Returns the primary monitor of the system.

Returns None if it can’t identify any monitor as a primary one.

pub fn monitor_from_point(&self, x: f64, y: f64) -> Result<Option<Monitor>>

Returns the monitor that contains the given point.

pub fn available_monitors(&self) -> Result<Vec<Monitor>>

Returns the list of all the monitors available on the system.

pub fn cursor_position(&self) -> Result<PhysicalPosition<f64>>

Get the cursor position relative to the top-left hand corner of the desktop.

Note that the top-left hand corner of the desktop is not necessarily the same as the screen. If the user uses a desktop with multiple monitors, the top-left hand corner of the desktop is the top-left hand corner of the main monitor on Windows and macOS or the top-left of the leftmost monitor on X11.

The coordinates can be negative if the top-left hand corner of the window is outside of the visible screen region.

pub fn set_theme(&self, theme: Option<Theme>)

Set the app theme.

pub fn default_window_icon(&self) -> Option<&Image<'_>>

Returns the default window icon.

pub fn menu(&self) -> Option<Menu<R>>

Returns the app-wide menu.

pub fn set_menu(&self, menu: Menu<R>) -> Result<Option<Menu<R>>>

Sets the app-wide menu and returns the previous one.

If a window was not created with an explicit menu or had one set explicitly, this menu will be assigned to it.

pub fn remove_menu(&self) -> Result<Option<Menu<R>>>

Remove the app-wide menu and returns it.

If a window was not created with an explicit menu or had one set explicitly, this will remove the menu from it.

pub fn hide_menu(&self) -> Result<()>

Hides the app-wide menu from windows that have it.

If a window was not created with an explicit menu or had one set explicitly, this will hide the menu from it.

pub fn show_menu(&self) -> Result<()>

Shows the app-wide menu for windows that have it.

If a window was not created with an explicit menu or had one set explicitly, this will show the menu for it.

pub fn show(&self) -> Result<()>

Shows the application, but does not automatically focus it.

pub fn hide(&self) -> Result<()>

Hides the application.

pub fn cleanup_before_exit(&self)

Runs necessary cleanup tasks before exiting the process. You should always exit the tauri app immediately after this function returns and not use any tauri-related APIs.

pub fn invoke_key(&self) -> &str

Gets the invoke key that must be referenced when using crate::webview::InvokeRequest.

Security

DO NOT expose this key to third party scripts as might grant access to the backend from external URLs and iframes.

impl<R: Runtime> App<R>

pub fn run_on_main_thread<F: FnOnce() + Send + 'static>( &self, f: F, ) -> Result<()>

Runs the given closure on the main thread.

pub fn handle(&self) -> &AppHandle<R>

Gets a handle to the application instance.

pub fn set_activation_policy(&mut self, activation_policy: ActivationPolicy)

Sets the activation policy for the application. It is set to NSApplicationActivationPolicyRegular by default.

Examples

tauri::Builder::default()
  .setup(move |app| {
    #[cfg(target_os = "macos")]
    app.set_activation_policy(tauri::ActivationPolicy::Accessory);
    Ok(())
  });

pub fn set_device_event_filter(&mut self, filter: DeviceEventFilter)

Change the device event filter mode.

Since the DeviceEvent capture can lead to high CPU usage for unfocused windows, tao will ignore them by default for unfocused windows on Windows. This method allows changing the filter to explicitly capture them again.

Platform-specific

** Linux / macOS / iOS / Android**: Unsupported.

Examples

let mut app = tauri::Builder::default()
  // on an actual app, remove the string argument
  .build(tauri::generate_context!("test/fixture/src-tauri/tauri.conf.json"))
  .expect("error while building tauri application");
app.set_device_event_filter(tauri::DeviceEventFilter::Always);
app.run(|_app_handle, _event| {});

pub fn run<F: FnMut(&AppHandle<R>, RunEvent) + 'static>(self, callback: F)

Runs the application.

Examples

let app = tauri::Builder::default()
  // on an actual app, remove the string argument
  .build(tauri::generate_context!("test/fixture/src-tauri/tauri.conf.json"))
  .expect("error while building tauri application");
app.run(|_app_handle, event| match event {
  tauri::RunEvent::ExitRequested { api, .. } => {
    api.prevent_exit();
  }
  _ => {}
});

pub fn run_iteration<F: FnMut(&AppHandle<R>, RunEvent) + 'static>( &mut self, callback: F, )

Runs an iteration of the runtime event loop and immediately return.

Note that when using this API, app cleanup is not automatically done. The cleanup calls App::cleanup_before_exit so you may want to call that function before exiting the application.

Examples

use tauri::Manager;

let mut app = tauri::Builder::default()
  // on an actual app, remove the string argument
  .build(tauri::generate_context!("test/fixture/src-tauri/tauri.conf.json"))
  .expect("error while building tauri application");

loop {
  app.run_iteration(|_app, _event| {});
  if app.webview_windows().is_empty() {
    app.cleanup_before_exit();
    break;
  }
}