Struct tauri::AppHandle

pub struct AppHandle<R: Runtime = Wry> { /* private fields */ }

A handle to the currently running application.

This type implements Manager which allows for manipulation of global application items.

Implementations

impl AppHandle<Wry>

APIs specific to the wry runtime.

pub fn create_tao_window<F: FnOnce() -> (String, WryWindowBuilder) + Send + 'static>( &self, f: F ) -> Result<Weak<Window>>

Create a new tao window using a callback. The event loop must be running at this point.

pub fn send_tao_window_event( &self, window_id: WindowId, message: WindowMessage ) -> Result<()>

Sends a window message to the event loop.

impl<R: Runtime> AppHandle<R>

pub fn run_on_main_thread<F: FnOnce() + Send + 'static>( &self, f: F ) -> Result<()>

Runs the given closure on the main thread.

pub fn plugin<P: Plugin<R> + 'static>(&self, plugin: P) -> Result<()>

Adds a Tauri application plugin. This function can be used to register a plugin that is loaded dynamically e.g. after login. For plugins that are created when the app is started, prefer Builder::plugin.

See Builder::plugin for more information.

Examples

use tauri::{plugin::{Builder as PluginBuilder, TauriPlugin}, Runtime};

fn init_plugin<R: Runtime>() -> TauriPlugin<R> {
  PluginBuilder::new("dummy").build()
}

tauri::Builder::default()
  .setup(move |app| {
    let handle = app.handle();
    std::thread::spawn(move || {
      handle.plugin(init_plugin());
    });

    Ok(())
  });

pub fn remove_plugin(&self, plugin: &'static str) -> bool

Removes the plugin with the given name.

Examples

use tauri::{plugin::{Builder as PluginBuilder, TauriPlugin, Plugin}, Runtime};

fn init_plugin<R: Runtime>() -> TauriPlugin<R> {
  PluginBuilder::new("dummy").build()
}

let plugin = init_plugin();
// `.name()` requires the `PLugin` trait import
let plugin_name = plugin.name();
tauri::Builder::default()
  .plugin(plugin)
  .setup(move |app| {
    let handle = app.handle();
    std::thread::spawn(move || {
      handle.remove_plugin(plugin_name);
    });

    Ok(())
  });

pub fn exit(&self, exit_code: i32)

Exits the app. This is the same as std::process::exit, but it performs cleanup on this application.

pub fn restart(&self)

Restarts the app. This is the same as crate::api::process::restart, but it performs cleanup on this application.

impl<R: Runtime> AppHandle<R>

pub fn path_resolver(&self) -> PathResolver

The path resolver for the application.

pub fn config(&self) -> Arc<Config>

Gets the app’s configuration, defined on the tauri.conf.json file.

pub fn package_info(&self) -> &PackageInfo

Gets the app’s package information.

pub fn asset_resolver(&self) -> AssetResolver<R>

The application’s asset resolver.

pub fn show(&self) -> Result<()>

Shows the application, but does not automatically focus it.

pub fn hide(&self) -> Result<()>

Hides the application.