Struct tauri::Context

pub struct Context<R: Runtime = Wry> {
    pub assets: Box<dyn Assets<R>>,
    /* private fields */
}

User supplied data required inside of a Tauri application.

Stability

This is the output of the generate_context macro, and is not considered part of the stable API. Unless you know what you are doing and are prepared for this type to have breaking changes, do not create it yourself.

Fields
assets: Box<dyn Assets<R>>

Asset provider.

Implementations

impl<R: Runtime> Context<R>

pub fn config(&self) -> &Config

The config the application was prepared with.

pub fn config_mut(&mut self) -> &mut Config

A mutable reference to the config the application was prepared with.

pub fn assets(&self) -> &dyn Assets<R>

The assets to be served directly by Tauri.

pub fn set_assets(&mut self, assets: Box<dyn Assets<R>>) -> Box<dyn Assets<R>>

Replace the Assets implementation and returns the previous value so you can use it as a fallback if desired.

pub fn default_window_icon(&self) -> Option<&Image<'_>>

The default window icon Tauri should use when creating windows.

pub fn set_default_window_icon(&mut self, icon: Option<Image<'static>>)

Set the default window icon Tauri should use when creating windows.

pub fn package_info(&self) -> &PackageInfo

Package information.

pub fn package_info_mut(&mut self) -> &mut PackageInfo

A mutable reference to the package information.

pub fn pattern(&self) -> &Pattern

The application pattern.

pub fn new( config: Config, assets: Box<dyn Assets<R>>, default_window_icon: Option<Image<'static>>, app_icon: Option<Vec<u8>>, package_info: PackageInfo, pattern: Pattern, runtime_authority: RuntimeAuthority, plugin_global_api_scripts: Option<&'static [&'static str]>, ) -> Self

Create a new Context from the minimal required items.