Struct tauri::ResourceTable

pub struct ResourceTable { /* private fields */ }

Map-like data structure storing Tauri’s resources (equivalent to file descriptors).

Provides basic methods for element access. A resource can be of any type. Different types of resources can be stored in the same map, and provided with a name for description.

Each resource is identified through a resource ID (rid), which acts as the key in the map.

Implementations

impl ResourceTable

pub fn add<T: Resource>(&mut self, resource: T) -> ResourceId

Inserts resource into the resource table, which takes ownership of it.

The resource type is erased at runtime and must be statically known when retrieving it through get().

Returns a unique resource ID, which acts as a key for this resource.

pub fn add_arc<T: Resource>(&mut self, resource: Arc<T>) -> ResourceId

Inserts a Arc-wrapped resource into the resource table.

The resource type is erased at runtime and must be statically known when retrieving it through get().

Returns a unique resource ID, which acts as a key for this resource.

pub fn add_arc_dyn(&mut self, resource: Arc<dyn Resource>) -> ResourceId

Inserts a Arc-wrapped resource into the resource table.

The resource type is erased at runtime and must be statically known when retrieving it through get().

Returns a unique resource ID, which acts as a key for this resource.

pub fn has(&self, rid: ResourceId) -> bool

Returns true if any resource with the given rid exists.

pub fn get<T: Resource>(&self, rid: ResourceId) -> Result<Arc<T>>

Returns a reference counted pointer to the resource of type T with the given rid. If rid is not present or has a type different than T, this function returns Error::BadResourceId.

pub fn get_any(&self, rid: ResourceId) -> Result<Arc<dyn Resource>>

Returns a reference counted pointer to the resource of the given rid. If rid is not present, this function returns [Error::BadResourceId].

pub fn replace<T: Resource>(&mut self, rid: ResourceId, resource: T)

Replaces a resource with a new resource.

Panics if the resource does not exist.

pub fn take<T: Resource>(&mut self, rid: ResourceId) -> Result<Arc<T>>

Removes a resource of type T from the resource table and returns it. If a resource with the given rid exists but its type does not match T, it is not removed from the resource table. Note that the resource’s close() method is not called.

Also note that there might be a case where the returned Arc<T> is referenced by other variables. That is, we cannot assume that Arc::strong_count(&returned_arc) is always equal to 1 on success. In particular, be really careful when you want to extract the inner value of type T from Arc<T>.

pub fn take_any(&mut self, rid: ResourceId) -> Result<Arc<dyn Resource>>

Removes a resource from the resource table and returns it. Note that the resource’s close() method is not called.

Also note that there might be a case where the returned Arc<T> is referenced by other variables. That is, we cannot assume that Arc::strong_count(&returned_arc) is always equal to 1 on success. In particular, be really careful when you want to extract the inner value of type T from Arc<T>.

pub fn names(&self) -> impl Iterator<Item = (ResourceId, Cow<'_, str>)>

Returns an iterator that yields a (id, name) pair for every resource that’s currently in the resource table. This can be used for debugging purposes. Note that the order in which items appear is not specified.

pub fn close(&mut self, rid: ResourceId) -> Result<()>

Removes the resource with the given rid from the resource table. If the only reference to this resource existed in the resource table, this will cause the resource to be dropped. However, since resources are reference counted, therefore pending ops are not automatically cancelled. A resource may implement the close() method to perform clean-ups such as canceling ops.