Struct tauri::WebviewAttributes

pub struct WebviewAttributes {

    pub url: WebviewUrl,
    pub user_agent: Option<String>,
    pub initialization_scripts: Vec<String>,
    pub data_directory: Option<PathBuf>,
    pub drag_drop_handler_enabled: bool,
    pub clipboard: bool,
    pub accept_first_mouse: bool,
    pub additional_browser_args: Option<String>,
    pub window_effects: Option<WindowEffectsConfig>,
    pub incognito: bool,
    pub transparent: bool,
    pub bounds: Option<Rect>,
    pub auto_resize: bool,
    pub proxy_url: Option<Url>,
    pub zoom_hotkeys_enabled: bool,
    pub browser_extensions_enabled: bool,

}

The attributes used to create an webview.

Fields
url: WebviewUrluser_agent: Option<String>initialization_scripts: Vec<String>data_directory: Option<PathBuf>drag_drop_handler_enabled: boolclipboard: boolaccept_first_mouse: booladditional_browser_args: Option<String>window_effects: Option<WindowEffectsConfig>incognito: booltransparent: boolbounds: Option<Rect>auto_resize: boolproxy_url: Option<Url>zoom_hotkeys_enabled: boolbrowser_extensions_enabled: bool
Implementations

impl WebviewAttributes

pub fn new(url: WebviewUrl) -> WebviewAttributes

Initializes the default attributes for a webview.

pub fn user_agent(self, user_agent: &str) -> WebviewAttributes

Sets the user agent

pub fn initialization_script(self, script: &str) -> WebviewAttributes

Sets the init script.

pub fn data_directory(self, data_directory: PathBuf) -> WebviewAttributes

Data directory for the webview.

pub fn disable_drag_drop_handler(self) -> WebviewAttributes

Disables the drag and drop handler. This is required to use HTML5 drag and drop APIs on the frontend on Windows.

pub fn enable_clipboard_access(self) -> WebviewAttributes

Enables clipboard access for the page rendered on Linux and Windows.

macOS doesnโt provide such method and is always enabled by default, but you still need to add menu item accelerators to use shortcuts.

pub fn accept_first_mouse(self, accept: bool) -> WebviewAttributes

Sets whether clicking an inactive window also clicks through to the webview.

pub fn additional_browser_args(self, additional_args: &str) -> WebviewAttributes

Sets additional browser arguments. Windows Only

pub fn window_effects(self, effects: WindowEffectsConfig) -> WebviewAttributes

Sets window effects

pub fn incognito(self, incognito: bool) -> WebviewAttributes

Enable or disable incognito mode for the WebView.

pub fn auto_resize(self) -> WebviewAttributes

Sets the webview to automatically grow and shrink its size and position when the parent window resizes.

pub fn proxy_url(self, url: Url) -> WebviewAttributes

Enable proxy for the WebView

pub fn zoom_hotkeys_enabled(self, enabled: bool) -> WebviewAttributes

Whether page zooming by hotkeys is enabled

Platform-specific:

Windows: Controls WebView2โs IsZoomControlEnabled setting.

MacOS / Linux: Injects a polyfill that zooms in and out with ctrl/command + -/=, 20% in each step, ranging from 20% to 1000%. Requires webview:allow-set-webview-zoom permission

Android / iOS: Unsupported.

pub fn browser_extensions_enabled(self, enabled: bool) -> WebviewAttributes

Whether browser extensions can be installed for the webview process

Platform-specific:

Windows: Enables the WebView2 environmentโs AreBrowserExtensionsEnabled

MacOS / Linux / iOS / Android - Unsupported.