Struct tauri::webview::Webview

pub struct Webview<R: Runtime = Wry> { /* private fields */ }

Webview.

Implementations

impl<R: Runtime> Webview<R>

Base webview functions.

pub fn run_on_main_thread<F: FnOnce() + Send + 'static>( &self, f: F, ) -> Result<()>

Runs the given closure on the main thread.

pub fn label(&self) -> &str

The webview label.

pub fn on_webview_event<F: Fn(&WebviewEvent) + Send + 'static>(&self, f: F)

Registers a window event listener.

impl<R: Runtime> Webview<R>

Desktop webview setters and actions.

pub fn print(&self) -> Result<()>

Opens the dialog to prints the contents of the webview. Currently only supported on macOS on wry. window.print() works on all platforms.

pub fn cursor_position(&self) -> Result<PhysicalPosition<f64>>

Get the cursor position relative to the top-left hand corner of the desktop.

Note that the top-left hand corner of the desktop is not necessarily the same as the screen. If the user uses a desktop with multiple monitors, the top-left hand corner of the desktop is the top-left hand corner of the main monitor on Windows and macOS or the top-left of the leftmost monitor on X11.

The coordinates can be negative if the top-left hand corner of the window is outside of the visible screen region.

pub fn close(&self) -> Result<()>

Closes this webview.

pub fn set_bounds(&self, bounds: Rect) -> Result<()>

Resizes this webview.

pub fn set_size<S: Into<Size>>(&self, size: S) -> Result<()>

Resizes this webview.

pub fn set_position<Pos: Into<Position>>(&self, position: Pos) -> Result<()>

Sets this webviews’s position.

pub fn set_focus(&self) -> Result<()>

Focus the webview.

pub fn hide(&self) -> Result<()>

Hide the webview.

pub fn show(&self) -> Result<()>

Show the webview.

pub fn reparent(&self, window: &Window<R>) -> Result<()>

Move the webview to the given window.

pub fn set_auto_resize(&self, auto_resize: bool) -> Result<()>

Sets whether the webview should automatically grow and shrink its size and position when the parent window resizes.

pub fn bounds(&self) -> Result<Rect>

Returns the bounds of the webviews’s client area.

pub fn position(&self) -> Result<PhysicalPosition<i32>>

Returns the webview position.

For child webviews, returns the position of the top-left hand corner of the webviews’s client area relative to the top-left hand corner of the parent window.

For webview window, returns the inner position of the window.

pub fn size(&self) -> Result<PhysicalSize<u32>>

Returns the physical size of the webviews’s client area.

impl<R: Runtime> Webview<R>

Webview APIs.

pub fn window(&self) -> Window<R>

The window that is hosting this webview.

pub fn window_ref(&self) -> MutexGuard<'_, Window<R>>

A reference to the window that is hosting this webview.

pub fn with_webview<F: FnOnce(PlatformWebview) + Send + 'static>( &self, f: F, ) -> Result<()>

Executes a closure, providing it with the webview handle that is specific to the current platform.

The closure is executed on the main thread.

Examples

pub fn url(&self) -> Result<Url>

Returns the current url of the webview.

pub fn navigate(&mut self, url: Url) -> Result<()>

Navigates the webview to the defined url.

pub fn on_message( self, request: InvokeRequest, responder: Box<OwnedInvokeResponder<R>>, )

Handles this window receiving an InvokeRequest.

pub fn eval(&self, js: &str) -> Result<()>

Evaluates JavaScript on this window.

pub fn open_devtools(&self)

Opens the developer tools window (Web Inspector). The devtools is only enabled on debug builds or with the devtools feature flag.

Platform-specific

macOS: Only supported on macOS 10.15+. This is a private API on macOS, so you cannot use this if your application will be published on the App Store.

Examples

pub fn close_devtools(&self)

Closes the developer tools window (Web Inspector). The devtools is only enabled on debug builds or with the devtools feature flag.

Platform-specific

macOS: Only supported on macOS 10.15+. This is a private API on macOS, so you cannot use this if your application will be published on the App Store.

Windows: Unsupported.

Examples

pub fn is_devtools_open(&self) -> bool

Checks if the developer tools window (Web Inspector) is opened. The devtools is only enabled on debug builds or with the devtools feature flag.

Platform-specific

macOS: Only supported on macOS 10.15+. This is a private API on macOS, so you cannot use this if your application will be published on the App Store.

Windows: Unsupported.

Examples

pub fn set_zoom(&self, scale_factor: f64) -> Result<()>

Set the webview zoom level

Platform-specific:

Android: Not supported.

macOS: available on macOS 11+ only.

iOS: available on iOS 14+ only.

pub fn clear_all_browsing_data(&self) -> Result<()>

Clear all browsing data for this webview.