Module tauri::api::path

Types and functions related to file system path operations.

Enums

BaseDirectory

A base directory to be used in resolve_path.

Functions

app_cache_dir

Returns the path to the suggested directory for your app’s cache files.

app_config_dir

Returns the path to the suggested directory for your app’s config files.

app_data_dir

Returns the path to the suggested directory for your app’s data files.

app_dirDeprecated

Returns the path to the suggested directory for your app’s config files.

app_local_data_dir

Returns the path to the suggested directory for your app’s local data files.

app_log_dir

Returns the path to the suggested directory for your app’s log files.

audio_dir

Returns the path to the user’s audio directory.

cache_dir

Returns the path to the user’s cache directory.

config_dir

Returns the path to the user’s config directory.

data_dir

Returns the path to the user’s data directory.

desktop_dir

Returns the path to the user’s desktop directory.

document_dir

Returns the path to the user’s document directory.

download_dir

Returns the path to the user’s download directory.

executable_dir

Returns the path to the user’s executable directory.

font_dir

Returns the path to the user’s font directory.

home_dir

Returns the path to the user’s home directory.

local_data_dir

Returns the path to the user’s local data directory.

log_dirDeprecated

Returns the path to the suggested directory for your app’s log files.

parse

Parse the given path, resolving a BaseDirectory variable if the path starts with one.

picture_dir

Returns the path to the user’s picture directory.

public_dir

Returns the path to the user’s public directory.

resolve_path

Resolves the path with the optional base directory.

resource_dir

Returns the path to the resource directory of this app.

runtime_dir

Returns the path to the user’s runtime directory.

template_dir

Returns the path to the user’s template directory.

video_dir

Returns the path to the user’s video dir