Module tauri::api::version

Compare two semantic versions.

Semantic Versioning is a guideline for how version numbers are assigned and incremented. The functions on this module are helpers around semver.

Functions

compare

Compare two semver versions.

is_compatible

Check if the “second” semver is compatible with the “first”.

is_greater

Check if a version is greater than the current.

is_major

Check if a the “other” version is a major bump from the “current”.

is_minor

Check if a the “other” version is a minor bump from the “current”.

is_patch

Check if a the “other” version is a patch bump from the “current”.