Trait tauri::command::CommandArg

pub trait CommandArg<'de, R: Runtime>: Sized {
    // Required method
    fn from_command(command: CommandItem<'de, R>) -> Result<Self, InvokeError>;
}

Trait implemented by command arguments to derive a value from a CommandItem.

Command Arguments

A command argument is any type that represents an item parsable from a CommandItem. Most implementations will use the data stored in InvokeMessage since CommandItem is mostly a wrapper around it.

Provided Implementations

Tauri implements CommandArg automatically for a number of types.

crate::Window

crate::State

T where T: serde::Deserialize

Any type that implements Deserialize can automatically be used as a CommandArg.

Required Methods

fn from_command(command: CommandItem<'de, R>) -> Result<Self, InvokeError>

Derives an instance of Self from the CommandItem.

If the derivation fails, the corresponding message will be rejected using InvokeMessage.