Enum tauri::MenuItem

#[non_exhaustive]
pub enum MenuItem {

    About(String, AboutMetadata),
    Hide,
    Services,
    HideOthers,
    ShowAll,
    CloseWindow,
    Quit,
    Copy,
    Cut,
    Undo,
    Redo,
    SelectAll,
    Paste,
    EnterFullScreen,
    Minimize,
    Zoom,
    Separator,

}

A menu item, bound to a pre-defined action or Custom emit an event. Note that status bar only supports Custom menu item variants. And on the menu bar, some platforms might not support some of the variants. Unsupported variant will be no-op on such platform.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

About(String, AboutMetadata)

Shows a standard “About” item.

The first value is the application name, and the second is its metadata.

Platform-specific

Windows / Android / iOS: Unsupported

Linux: The metadata is only applied on Linux

Hide

A standard “hide the app” menu item.

Platform-specific

Android / iOS: Unsupported

Services

A standard “Services” menu item.

Platform-specific

Windows / Linux / Android / iOS: Unsupported

HideOthers

A “hide all other windows” menu item.

Platform-specific

Windows / Linux / Android / iOS: Unsupported

ShowAll

A menu item to show all the windows for this app.

Platform-specific

Windows / Linux / Android / iOS: Unsupported

CloseWindow

Close the current window.

Platform-specific

Android / iOS: Unsupported

Quit

A “quit this app” menu icon.

Platform-specific

Android / iOS: Unsupported

Copy

A menu item for enabling copying (often text) from responders.

Platform-specific

Android / iOS / Linux: Unsupported

Cut

A menu item for enabling cutting (often text) from responders.

Platform-specific

Android / iOS / Linux: Unsupported

Undo

An “undo” menu item; particularly useful for supporting the cut/copy/paste/undo lifecycle of events.

Platform-specific

Windows / Linux / Android / iOS: Unsupported

Redo

An “redo” menu item; particularly useful for supporting the cut/copy/paste/undo lifecycle of events.

Platform-specific

Windows / Linux / Android / iOS: Unsupported

SelectAll

A menu item for selecting all (often text) from responders.

Platform-specific

Windows / Android / iOS / Linux: Unsupported

Paste

A menu item for pasting (often text) into responders.

Platform-specific

Android / iOS / Linux: Unsupported

EnterFullScreen

A standard “enter full screen” item.

Platform-specific

Windows / Linux / Android / iOS: Unsupported

Minimize

An item for minimizing the window with the standard system controls.

Platform-specific

Android / iOS: Unsupported

Zoom

An item for instructing the app to zoom

Platform-specific

Windows / Linux / Android / iOS: Unsupported

Separator

Represents a Separator

Platform-specific

Android / iOS: Unsupported