Macro tauri::tauri_build_context

macro_rules! tauri_build_context {
    () => { ... };
}

Include a Context that was generated by tauri-build inside your build script.

You should either use tauri-build and this macro to include the compile time generated code, or generate_context!. Do not use both at the same time, as they generate the same code and will cause excess computations that will be discarded.