Struct tauri::App

pub struct App<R: Runtime = Wry> { /* private fields */ }

The instance of the currently running application.

This type implements Manager which allows for manipulation of global application items.

Implementations

impl App<Wry>

APIs specific to the wry runtime.

pub fn wry_plugin<P: PluginBuilder<EventLoopMessage> + 'static>( &mut self, plugin: P )

Adds a tauri_runtime_wry::Plugin using its tauri_runtime_wry::PluginBuilder.

Stability

This API is unstable.

impl<R: Runtime> App<R>

pub fn path_resolver(&self) -> PathResolver

The path resolver for the application.

pub fn config(&self) -> Arc<Config>

Gets the app’s configuration, defined on the tauri.conf.json file.

pub fn package_info(&self) -> &PackageInfo

Gets the app’s package information.

pub fn asset_resolver(&self) -> AssetResolver<R>

The application’s asset resolver.

pub fn show(&self) -> Result<()>

Shows the application, but does not automatically focus it.

pub fn hide(&self) -> Result<()>

Hides the application.

impl<R: Runtime> App<R>

pub fn handle(&self) -> AppHandle<R>

Gets a handle to the application instance.

pub fn set_activation_policy(&mut self, activation_policy: ActivationPolicy)

Sets the activation policy for the application. It is set to NSApplicationActivationPolicyRegular by default.

Examples

let mut app = tauri::Builder::default()
  // on an actual app, remove the string argument
  .build(tauri::generate_context!("test/fixture/src-tauri/tauri.conf.json"))
  .expect("error while building tauri application");
#[cfg(target_os = "macos")]
app.set_activation_policy(tauri::ActivationPolicy::Accessory);
app.run(|_app_handle, _event| {});

pub fn set_device_event_filter(&mut self, filter: DeviceEventFilter)

Change the device event filter mode.

Since the DeviceEvent capture can lead to high CPU usage for unfocused windows, tao will ignore them by default for unfocused windows on Windows. This method allows changing the filter to explicitly capture them again.

Platform-specific

** Linux / macOS / iOS / Android**: Unsupported.

Examples

let mut app = tauri::Builder::default()
  // on an actual app, remove the string argument
  .build(tauri::generate_context!("test/fixture/src-tauri/tauri.conf.json"))
  .expect("error while building tauri application");
app.set_device_event_filter(tauri::DeviceEventFilter::Always);
app.run(|_app_handle, _event| {});

pub fn run<F: FnMut(&AppHandle<R>, RunEvent) + 'static>(self, callback: F)

Runs the application.

Examples

let app = tauri::Builder::default()
  // on an actual app, remove the string argument
  .build(tauri::generate_context!("test/fixture/src-tauri/tauri.conf.json"))
  .expect("error while building tauri application");
app.run(|_app_handle, event| match event {
  tauri::RunEvent::ExitRequested { api, .. } => {
    api.prevent_exit();
  }
  _ => {}
});

pub fn run_iteration(&mut self) -> RunIteration

Runs a iteration of the runtime event loop and immediately return.

Note that when using this API, app cleanup is not automatically done. The cleanup calls [crate::api::process::kill_children] so you may want to call that function before exiting the application. Additionally, the cleanup calls AppHandle#remove_system_tray (Windows only).

Examples

let mut app = tauri::Builder::default()
  // on an actual app, remove the string argument
  .build(tauri::generate_context!("test/fixture/src-tauri/tauri.conf.json"))
  .expect("error while building tauri application");
loop {
  let iteration = app.run_iteration();
  if iteration.window_count == 0 {
    break;
  }
}