Struct tauri::Context

pub struct Context<A: Assets> { /* private fields */ }

User supplied data required inside of a Tauri application.

Stability

This is the output of the generate_context macro, and is not considered part of the stable API. Unless you know what you are doing and are prepared for this type to have breaking changes, do not create it yourself.

Implementations

impl<A: Assets> Context<A>

pub fn config(&self) -> &Config

The config the application was prepared with.

pub fn config_mut(&mut self) -> &mut Config

A mutable reference to the config the application was prepared with.

pub fn assets(&self) -> Arc<A>

The assets to be served directly by Tauri.

pub fn assets_mut(&mut self) -> &mut Arc<A>

A mutable reference to the assets to be served directly by Tauri.

pub fn default_window_icon(&self) -> Option<&Icon>

The default window icon Tauri should use when creating windows.

pub fn default_window_icon_mut(&mut self) -> &mut Option<Icon>

A mutable reference to the default window icon Tauri should use when creating windows.

pub fn system_tray_icon(&self) -> Option<&Icon>

The icon to use on the system tray UI.

pub fn system_tray_icon_mut(&mut self) -> &mut Option<Icon>

A mutable reference to the icon to use on the system tray UI.

pub fn package_info(&self) -> &PackageInfo

Package information.

pub fn package_info_mut(&mut self) -> &mut PackageInfo

A mutable reference to the package information.

pub fn pattern(&self) -> &Pattern

The application pattern.

pub fn new( config: Config, assets: Arc<A>, default_window_icon: Option<Icon>, app_icon: Option<Vec<u8>>, system_tray_icon: Option<Icon>, package_info: PackageInfo, info_plist: (), pattern: Pattern ) -> Self

Create a new Context from the minimal required items.