Struct tauri::window::Window

pub struct Window<R: Runtime = Wry> { /* private fields */ }

A window managed by Tauri.

This type also implements Manager which allows you to manage other windows attached to the same application.

Implementations

impl<R: Runtime> Window<R>

Base window functions.

pub fn webviews(&self) -> Vec<Webview<R>>

List of webviews associated with this window.

pub fn run_on_main_thread<F: FnOnce() + Send + 'static>( &self, f: F, ) -> Result<()>

Runs the given closure on the main thread.

pub fn label(&self) -> &str

The label of this window.

pub fn on_window_event<F: Fn(&WindowEvent) + Send + 'static>(&self, f: F)

Registers a window event listener.

impl<R: Runtime> Window<R>

Menu APIs

pub fn on_menu_event<F: Fn(&Window<R>, MenuEvent) + Send + Sync + 'static>( &self, f: F, )

Registers a global menu event listener.

Note that this handler is called for any menu event, whether it is coming from this window, another window or from the tray icon menu.

Also note that this handler will not be called if the window used to register it was closed.

Examples

pub fn menu(&self) -> Option<Menu<R>>

Returns this window menu .

pub fn set_menu(&self, menu: Menu<R>) -> Result<Option<Menu<R>>>

Sets the window menu and returns the previous one.

Platform-specific:

macOS: Unsupported. The menu on macOS is app-wide and not specific to one window, if you need to set it, use AppHandle::set_menu instead.

pub fn remove_menu(&self) -> Result<Option<Menu<R>>>

Removes the window menu and returns it.

Platform-specific:

macOS: Unsupported. The menu on macOS is app-wide and not specific to one window, if you need to remove it, use AppHandle::remove_menu instead.

pub fn hide_menu(&self) -> Result<()>

Hides the window menu.

pub fn show_menu(&self) -> Result<()>

Shows the window menu.

pub fn is_menu_visible(&self) -> Result<bool>

Shows the window menu.

pub fn popup_menu<M: ContextMenu>(&self, menu: &M) -> Result<()>

Shows the specified menu as a context menu at the cursor position.

pub fn popup_menu_at<M: ContextMenu, P: Into<Position>>( &self, menu: &M, position: P, ) -> Result<()>

Shows the specified menu as a context menu at the specified position.

The position is relative to the window’s top-left corner.

impl<R: Runtime> Window<R>

Window getters.

pub fn scale_factor(&self) -> Result<f64>

Returns the scale factor that can be used to map logical pixels to physical pixels, and vice versa.

pub fn inner_position(&self) -> Result<PhysicalPosition<i32>>

Returns the position of the top-left hand corner of the window’s client area relative to the top-left hand corner of the desktop.

pub fn outer_position(&self) -> Result<PhysicalPosition<i32>>

Returns the position of the top-left hand corner of the window relative to the top-left hand corner of the desktop.

pub fn inner_size(&self) -> Result<PhysicalSize<u32>>

Returns the physical size of the window’s client area.

The client area is the content of the window, excluding the title bar and borders.

pub fn outer_size(&self) -> Result<PhysicalSize<u32>>

Returns the physical size of the entire window.

These dimensions include the title bar and borders. If you don’t want that (and you usually don’t), use inner_size instead.

pub fn is_fullscreen(&self) -> Result<bool>

Gets the window’s current fullscreen state.

pub fn is_minimized(&self) -> Result<bool>

Gets the window’s current minimized state.

pub fn is_maximized(&self) -> Result<bool>

Gets the window’s current maximized state.

pub fn is_focused(&self) -> Result<bool>

Gets the window’s current focus state.

pub fn is_decorated(&self) -> Result<bool>

Gets the window’s current decoration state.

pub fn is_resizable(&self) -> Result<bool>

Gets the window’s current resizable state.

pub fn is_enabled(&self) -> Result<bool>

Whether the window is enabled or disabled.

pub fn is_maximizable(&self) -> Result<bool>

Gets the window’s native maximize button state

Platform-specific

Linux / iOS / Android: Unsupported.

pub fn is_minimizable(&self) -> Result<bool>

Gets the window’s native minimize button state

Platform-specific

Linux / iOS / Android: Unsupported.

pub fn is_closable(&self) -> Result<bool>

Gets the window’s native close button state

Platform-specific

Linux / iOS / Android: Unsupported.

pub fn is_visible(&self) -> Result<bool>

Gets the window’s current visibility state.

pub fn title(&self) -> Result<String>

Gets the window’s current title.

pub fn current_monitor(&self) -> Result<Option<Monitor>>

Returns the monitor on which the window currently resides.

Returns None if current monitor can’t be detected.

pub fn monitor_from_point(&self, x: f64, y: f64) -> Result<Option<Monitor>>

Returns the monitor that contains the given point.

pub fn primary_monitor(&self) -> Result<Option<Monitor>>

Returns the primary monitor of the system.

Returns None if it can’t identify any monitor as a primary one.

pub fn available_monitors(&self) -> Result<Vec<Monitor>>

Returns the list of all the monitors available on the system.

pub fn ns_window(&self) -> Result<*mut c_void>

Returns the native handle that is used by this window.

pub fn ns_view(&self) -> Result<*mut c_void>

Returns the pointer to the content view of this window.

pub fn theme(&self) -> Result<Theme>

Returns the current window theme.

Platform-specific

macOS: Only supported on macOS 10.14+.

impl<R: Runtime> Window<R>

Desktop window getters.

pub fn cursor_position(&self) -> Result<PhysicalPosition<f64>>

Get the cursor position relative to the top-left hand corner of the desktop.

Note that the top-left hand corner of the desktop is not necessarily the same as the screen. If the user uses a desktop with multiple monitors, the top-left hand corner of the desktop is the top-left hand corner of the main monitor on Windows and macOS or the top-left of the leftmost monitor on X11.

The coordinates can be negative if the top-left hand corner of the window is outside of the visible screen region.

impl<R: Runtime> Window<R>

Desktop window setters and actions.

pub fn center(&self) -> Result<()>

Centers the window.

pub fn request_user_attention( &self, request_type: Option<UserAttentionType>, ) -> Result<()>

Requests user attention to the window, this has no effect if the application is already focused. How requesting for user attention manifests is platform dependent, see UserAttentionType for details.

Providing None will unset the request for user attention. Unsetting the request for user attention might not be done automatically by the WM when the window receives input.

Platform-specific

macOS: None has no effect.

Linux: Urgency levels have the same effect.

pub fn set_resizable(&self, resizable: bool) -> Result<()>

Determines if this window should be resizable. When resizable is set to false, native window’s maximize button is automatically disabled.

pub fn set_maximizable(&self, maximizable: bool) -> Result<()>

Determines if this window’s native maximize button should be enabled. If resizable is set to false, this setting is ignored.

Platform-specific

macOS: Disables the “zoom” button in the window titlebar, which is also used to enter fullscreen mode.

Linux / iOS / Android: Unsupported.

pub fn set_minimizable(&self, minimizable: bool) -> Result<()>

Determines if this window’s native minimize button should be enabled.

Platform-specific

Linux / iOS / Android: Unsupported.

pub fn set_closable(&self, closable: bool) -> Result<()>

Determines if this window’s native close button should be enabled.

Platform-specific

Linux: “GTK+ will do its best to convince the window manager not to show a close button. Depending on the system, this function may not have any effect when called on a window that is already visible”

iOS / Android: Unsupported.

pub fn set_title(&self, title: &str) -> Result<()>

Set this window’s title.

pub fn set_enabled(&self, enabled: bool) -> Result<()>

Enable or disable the window.

pub fn maximize(&self) -> Result<()>

Maximizes this window.

pub fn unmaximize(&self) -> Result<()>

Un-maximizes this window.

pub fn minimize(&self) -> Result<()>

Minimizes this window.

pub fn unminimize(&self) -> Result<()>

Un-minimizes this window.

pub fn show(&self) -> Result<()>

Show this window.

pub fn hide(&self) -> Result<()>

Hide this window.

pub fn close(&self) -> Result<()>

Closes this window. It emits [crate::RunEvent::CloseRequested] first like a user-initiated close request so you can intercept it.

pub fn destroy(&self) -> Result<()>

Destroys this window. Similar to Self::close but does not emit any events and force close the window instead.

pub fn set_decorations(&self, decorations: bool) -> Result<()>

Determines if this window should be decorated.

pub fn set_shadow(&self, enable: bool) -> Result<()>

Determines if this window should have shadow.

Platform-specific

Windows:

false has no effect on decorated window, shadow are always ON.

true will make undecorated window have a 1px white border, and on Windows 11, it will have a rounded corners.

Linux: Unsupported.

pub fn set_effects<E: Into<Option<WindowEffectsConfig>>>( &self, effects: E, ) -> Result<()>

Sets window effects, pass None to clear any effects applied if possible.

Requires the window to be transparent.

See EffectsBuilder for a convenient builder for WindowEffectsConfig.

Platform-specific:

Windows: If using decorations or shadows, you may want to try this workaround https://github.com/tauri-apps/tao/issues/72#issuecomment-975607891

Linux: Unsupported

pub fn set_always_on_bottom(&self, always_on_bottom: bool) -> Result<()>

Determines if this window should always be below other windows.

pub fn set_always_on_top(&self, always_on_top: bool) -> Result<()>

Determines if this window should always be on top of other windows.

pub fn set_visible_on_all_workspaces( &self, visible_on_all_workspaces: bool, ) -> Result<()>

Sets whether the window should be visible on all workspaces or virtual desktops.

Platform-specific

Windows / iOS / Android: Unsupported.

pub fn set_content_protected(&self, protected: bool) -> Result<()>

Prevents the window contents from being captured by other apps.

pub fn set_size<S: Into<Size>>(&self, size: S) -> Result<()>

Resizes this window.

pub fn set_min_size<S: Into<Size>>(&self, size: Option<S>) -> Result<()>

Sets this window’s minimum inner size.

pub fn set_max_size<S: Into<Size>>(&self, size: Option<S>) -> Result<()>

Sets this window’s maximum inner size.

pub fn set_size_constraints( &self, constriants: WindowSizeConstraints, ) -> Result<()>

Sets this window’s minimum inner width.

pub fn set_position<Pos: Into<Position>>(&self, position: Pos) -> Result<()>

Sets this window’s position.

pub fn set_fullscreen(&self, fullscreen: bool) -> Result<()>

Determines if this window should be fullscreen.

pub fn set_focus(&self) -> Result<()>

Bring the window to front and focus.

pub fn set_icon(&self, icon: Image<'_>) -> Result<()>

Sets this window’ icon.

pub fn set_skip_taskbar(&self, skip: bool) -> Result<()>

Whether to hide the window icon from the taskbar or not.

Platform-specific

macOS: Unsupported.

pub fn set_cursor_grab(&self, grab: bool) -> Result<()>

Grabs the cursor, preventing it from leaving the window.

There’s no guarantee that the cursor will be hidden. You should hide it by yourself if you want so.

Platform-specific

Linux: Unsupported.

macOS: This locks the cursor in a fixed location, which looks visually awkward.

pub fn set_cursor_visible(&self, visible: bool) -> Result<()>

Modifies the cursor’s visibility.

If false, this will hide the cursor. If true, this will show the cursor.

Platform-specific

Windows: The cursor is only hidden within the confines of the window.

macOS: The cursor is hidden as long as the window has input focus, even if the cursor is outside of the window.

pub fn set_cursor_icon(&self, icon: CursorIcon) -> Result<()>

Modifies the cursor icon of the window.

pub fn set_cursor_position<Pos: Into<Position>>( &self, position: Pos, ) -> Result<()>

Changes the position of the cursor in window coordinates.

pub fn set_ignore_cursor_events(&self, ignore: bool) -> Result<()>

Ignores the window cursor events.

pub fn start_dragging(&self) -> Result<()>

Starts dragging the window.

pub fn start_resize_dragging(&self, direction: ResizeDirection) -> Result<()>

Starts resize-dragging the window.

pub fn set_progress_bar(&self, progress_state: ProgressBarState) -> Result<()>

Sets the taskbar progress state.

Platform-specific

Linux / macOS: Progress bar is app-wide and not specific to this window.

Linux: Only supported desktop environments with libunity (e.g. GNOME).

iOS / Android: Unsupported.

pub fn set_title_bar_style(&self, style: TitleBarStyle) -> Result<()>

Sets the title bar style. macOS only.

pub fn set_theme(&self, theme: Option<Theme>) -> Result<()>

Sets the theme for this window.

Platform-specific

Linux / macOS: Theme is app-wide and not specific to this window.

iOS / Android: Unsupported.